"long" audit. can't use long when we want 32 bits. May be a few more cases.

+++ include/QF/GL/extensions.h
@@ -311,7 +311,7 @@ typedef void (GLAPIENTRY *QF_glTexCoordPointerEXT) (GLint size, GLenum type, GLs
 typedef void (GLAPIENTRY *QF_glVertexPointerEXT) (GLint size, GLenum type, GLsizei stride, GLsizei count, const GLvoid *pointer);
 
 // 3Dfx
-typedef long    FxI32;
+typedef int     FxI32;
 typedef FxI32   GrScreenResolution_t;
 typedef FxI32   GrDitherMode_t;
 typedef FxI32   GrScreenRefresh_t;
